Add kl scaling for params, fix dtype issue

+++ dreamer/blocks.py
@@ -126,7 +126,7 @@ class MeanField(hk.Module):
         # Add a bias to softplus such that a zero value to the stddevs parameter
         # gives the empirical standard deviation.
         empirical_stddev = flat_params.std()
-        init = np.log(np.exp(empirical_stddev) - 1.0)
+        init = np.log(np.exp(empirical_stddev) - 1.0).astype(mus.dtype)
         stddevs = jnn.softplus(stddevs + init) + 1e-6
         return tfd.MultivariateNormalDiag(mus, stddevs)
